+++ example/lib/search_feed.dart
@@ -1,8 +1,6 @@
 import 'package:example/data_model.dart';
 import 'package:firestore_search/firestore_search.dart';
 import 'package:flutter/material.dart';
-export 'package:firestore_search/src/firestore_service.dart'
-    hide FirestoreService;
 
 class SearchFeed extends StatefulWidget {
   const SearchFeed({Key? key}) : super(key: key);
@@ -17,7 +15,80 @@ class _SearchFeedState extends State<SearchFeed> {
     return FirestoreSearchScaffold(
       firestoreCollectionName: 'packages',
       searchBy: 'tool',
-      scaffoldBody: const Center(child: Text('Firestore Search')),
+      scaffoldBody: Column(
+        children: [
+          Row(
+            children: [
+              const SizedBox(
+                width: 40.0,
+              ),
+              Expanded(
+                  child: FirestoreSearchBar(
+                tag: 'test',
+              )),
+              const SizedBox(
+                width: 40.0,
+              ),
+            ],
+          ),
+          Expanded(
+            child: FirestoreSearchResults.builder(
+              tag: 'test',
+              firestoreCollectionName: 'packages',
+              searchBy: 'tool',
+              initialBody: const Center(child: Text('Initial body'),),
+              dataListFromSnapshot: DataModel().dataListFromSnapshot,
+              builder: (context, snapshot) {
+                if (snapshot.hasData) {
+                  final List<DataModel>? dataList = snapshot.data;
+                  if (dataList!.isEmpty) {
+                    return const Center(
+                      child: Text('No Results Returned'),
+                    );
+                  }
+                  return ListView.builder(
+                      itemCount: dataList.length,
+                      itemBuilder: (context, index) {
+                        final DataModel data = dataList[index];
+
+                        return Column(
+                          mainAxisSize: MainAxisSize.min,
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            Padding(
+                              padding: const EdgeInsets.all(8.0),
+                              child: Text(
+                                '${data.name}',
+                                style: Theme.of(context).textTheme.headline6,
+                              ),
+                            ),
+                            Padding(
+                              padding: const EdgeInsets.only(
+                                  bottom: 8.0, left: 8.0, right: 8.0),
+                              child: Text('${data.developer}',
+                                  style: Theme.of(context).textTheme.bodyText1),
+                            )
+                          ],
+                        );
+                      });
+                }
+
+                if (snapshot.connectionState == ConnectionState.done) {
+                  if (!snapshot.hasData) {
+                    return const Center(
+                      child: Text('No Results Returned'),
+                    );
+                  }
+                }
+                return const Center(
+                  child: CircularProgressIndicator(),
+                );
+              },
+            ),
+          )
+        ],
+      ),
       dataListFromSnapshot: DataModel().dataListFromSnapshot,
       builder: (context, snapshot) {
         if (snapshot.hasData) {